requirements.txt
Flask
python-whois